requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
motor
requests
python-dotenv
supabase
pinecone-client
python-dotenv
langchain_community
langchain
langchain_google_genai
pinecone_text